backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
openai==1.3.0
pandas==2.1.3
openpyxl==3.1.2
scikit-learn==1.3.2
python-dotenv==1.0.0